Handle Firmatas, move inquirer prompt to lib
