+++ lib/interchange.js
@@ -353,7 +353,7 @@ Interchange.prototype.check_firmware = function (firmware, options, cb) {
     // to a temporary location
 
     var board = options.board || "nano"; // assumes nano if none provided
-    var useFirmata = (firmware.indexOf('Firmata') > 0) || (options.firmata != null) || false;
+    var useFirmata = (firmware.indexOf('Firmata') > 0) || options.firmata || false;
     var firmataName = options.firmata || "";
     // check for default where no firmata name is supplied or default is implied.
     if (firmataName === true) {
